Refactors OpenMM protocols into own module.
