+++ propertyestimator/tests/test_protocols/test_defaults.py
@@ -28,7 +28,16 @@ for protocol_module in protocol_modules:
 def test_default_protocol_schemas(available_protocol):
     """A simple test to ensure that each available protocol
     can both create, and be created from a schema."""
-    protocol = registered_workflow_protocols[available_protocol]("dummy_id")
+    protocol_class = registered_workflow_protocols[available_protocol]
+
+    if (
+        protocol_class.__abstractmethods__ is not None
+        and len(protocol_class.__abstractmethods__) > 0
+    ):
+        # Skip base classes.
+        return
+
+    protocol = protocol_class("dummy_id")
     protocol_schema = protocol.schema
 
     recreated_protocol = registered_workflow_protocols[available_protocol]("dummy_id")
