+++ propertyestimator/tests/test_protocols/test_gradient_protocols.py
@@ -1,57 +1,10 @@
 import random
 import tempfile
-from os import path
-
-import pytest
 
 from propertyestimator import unit
 from propertyestimator.backends import ComputeResources
 from propertyestimator.forcefield import ParameterGradientKey
-from propertyestimator.protocols.gradients import (
-    CentralDifferenceGradient,
-    GradientReducedPotentials,
-)
-from propertyestimator.substances import Substance
-from propertyestimator.tests.utils import build_tip3p_smirnoff_force_field
-from propertyestimator.thermodynamics import ThermodynamicState
-from propertyestimator.utils import get_data_filename
-
-
-@pytest.mark.parametrize("use_subset", [True, False])
-def test_gradient_reduced_potentials(use_subset):
-
-    substance = Substance.from_components("O")
-    thermodynamic_state = ThermodynamicState(298 * unit.kelvin, 1.0 * unit.atmosphere)
-
-    with tempfile.TemporaryDirectory() as directory:
-
-        force_field_path = path.join(directory, "ff.json")
-
-        with open(force_field_path, "w") as file:
-            file.write(build_tip3p_smirnoff_force_field().json())
-
-        reduced_potentials = GradientReducedPotentials(f"reduced_potentials")
-        reduced_potentials.substance = substance
-        reduced_potentials.thermodynamic_state = thermodynamic_state
-        reduced_potentials.statistics_path = get_data_filename(
-            "test/statistics/stats_pandas.csv"
-        )
-        reduced_potentials.force_field_path = force_field_path
-        reduced_potentials.trajectory_file_path = get_data_filename(
-            "test/trajectories/water.dcd"
-        )
-        reduced_potentials.coordinate_file_path = get_data_filename(
-            "test/trajectories/water.pdb"
-        )
-        reduced_potentials.use_subset_of_force_field = use_subset
-        reduced_potentials.enable_pbc = True
-        reduced_potentials.parameter_key = ParameterGradientKey(
-            "vdW", "[#1]-[#8X2H2+0:1]-[#1]", "epsilon"
-        )
-
-        reduced_potentials.execute(directory, ComputeResources())
-        assert path.isfile(reduced_potentials.forward_potentials_path)
-        assert path.isfile(reduced_potentials.reverse_potentials_path)
+from propertyestimator.protocols.gradients import CentralDifferenceGradient
 
 
 def test_central_difference_gradient():
