+++ propertyestimator/tests/test_protocols/test_openmm.py
@@ -1,22 +1,27 @@
 """
-Units tests for propertyestimator.protocols.simulation
+Units tests for propertyestimator.protocols.openmm
 """
 import json
 import os
 import tempfile
 from os import path
 
+import pytest
+
 from propertyestimator import unit
 from propertyestimator.backends import ComputeResources
+from propertyestimator.forcefield import ParameterGradientKey
 from propertyestimator.protocols.coordinates import BuildCoordinatesPackmol
 from propertyestimator.protocols.forcefield import BuildSmirnoffSystem
-from propertyestimator.protocols.simulation import (
-    RunEnergyMinimisation,
-    RunOpenMMSimulation,
+from propertyestimator.protocols.openmm import (
+    OpenMMEnergyMinimisation,
+    OpenMMGradientPotentials,
+    OpenMMSimulation,
 )
 from propertyestimator.substances import Substance
 from propertyestimator.tests.utils import build_tip3p_smirnoff_force_field
 from propertyestimator.thermodynamics import ThermodynamicState
+from propertyestimator.utils import get_data_filename
 from propertyestimator.utils.serialization import TypedJSONDecoder, TypedJSONEncoder
 from propertyestimator.utils.statistics import StatisticsArray
 
@@ -51,7 +56,7 @@ def test_run_energy_minimisation():
 
         coordinate_path, system_path = _setup_dummy_system(directory)
 
-        energy_minimisation = RunEnergyMinimisation("energy_minimisation")
+        energy_minimisation = OpenMMEnergyMinimisation("energy_minimisation")
         energy_minimisation.input_coordinate_file = coordinate_path
         energy_minimisation.system_path = system_path
         energy_minimisation.execute(directory, ComputeResources())
@@ -66,7 +71,7 @@ def test_run_openmm_simulation():
 
         coordinate_path, system_path = _setup_dummy_system(directory)
 
-        npt_equilibration = RunOpenMMSimulation("npt_equilibration")
+        npt_equilibration = OpenMMSimulation("npt_equilibration")
         npt_equilibration.steps_per_iteration = 2
         npt_equilibration.output_frequency = 1
         npt_equilibration.thermodynamic_state = thermodynamic_state
@@ -90,7 +95,7 @@ def test_run_openmm_simulation_checkpoints():
         coordinate_path, system_path = _setup_dummy_system(directory)
 
         # Check that executing twice doesn't run the simulation twice
-        npt_equilibration = RunOpenMMSimulation("npt_equilibration")
+        npt_equilibration = OpenMMSimulation("npt_equilibration")
         npt_equilibration.total_number_of_iterations = 1
         npt_equilibration.steps_per_iteration = 4
         npt_equilibration.output_frequency = 1
@@ -139,3 +144,40 @@ def test_run_openmm_simulation_checkpoints():
             )
             == 4
         )
+
+
+@pytest.mark.parametrize("use_subset", [True, False])
+def test_gradient_reduced_potentials(use_subset):
+
+    substance = Substance.from_components("O")
+    thermodynamic_state = ThermodynamicState(298 * unit.kelvin, 1.0 * unit.atmosphere)
+
+    with tempfile.TemporaryDirectory() as directory:
+
+        force_field_path = path.join(directory, "ff.json")
+
+        with open(force_field_path, "w") as file:
+            file.write(build_tip3p_smirnoff_force_field().json())
+
+        reduced_potentials = OpenMMGradientPotentials(f"reduced_potentials")
+        reduced_potentials.substance = substance
+        reduced_potentials.thermodynamic_state = thermodynamic_state
+        reduced_potentials.statistics_path = get_data_filename(
+            "test/statistics/stats_pandas.csv"
+        )
+        reduced_potentials.force_field_path = force_field_path
+        reduced_potentials.trajectory_file_path = get_data_filename(
+            "test/trajectories/water.dcd"
+        )
+        reduced_potentials.coordinate_file_path = get_data_filename(
+            "test/trajectories/water.pdb"
+        )
+        reduced_potentials.use_subset_of_force_field = use_subset
+        reduced_potentials.enable_pbc = True
+        reduced_potentials.parameter_key = ParameterGradientKey(
+            "vdW", "[#1]-[#8X2H2+0:1]-[#1]", "epsilon"
+        )
+
+        reduced_potentials.execute(directory, ComputeResources())
+        assert path.isfile(reduced_potentials.forward_potentials_path)
+        assert path.isfile(reduced_potentials.reverse_potentials_path)
